♻️ refactor(reasoning): migrate reasoning mode management to registry
- replace prompts.py with registry.py for reasoning mode creation and discovery
- update coordinator to use registry functions for reasoning mode instantiation
- streamline reasoning mode initialization and logging messages for clarity
